move test to functional
